AlertDetails: Add missing fields
